+++ decomposition/include/Sub_Matrix_Decomposition_Cost_Function.h
@@ -67,7 +67,7 @@ functor_extract_submatrices( Matrix& matrix_in, std::vector<Matrix, tbb::cache_a
 @brief Operator to extract the sumbatrix indexed by submtx_idx
 @param r A range of indices labeling the given submatrix to be extracted
 */
-void operator()( tbb::blocked_range<size_t> r ) const;
+void operator()( tbb::blocked_range<int> r ) const;
 
 };
 
@@ -97,7 +97,7 @@ public:
 @param prod_num_in The number of partial cost function values (equal to the number of distinct submatrix products.)
 @return Returns with the instance of the class.
 */
-functor_submtx_cost_fnc( std::vector<Matrix, tbb::cache_aligned_allocator<Matrix>>* submatrices_in, tbb::combinable<double>* prod_cost_functions_in, size_t prod_num_in );
+functor_submtx_cost_fnc( std::vector<Matrix, tbb::cache_aligned_allocator<Matrix>>* submatrices_in, tbb::combinable<double>* prod_cost_functions_in, int prod_num_in );
 
 /**
 @brief Operator to calculate the partial cost function labeled by product_idx
